backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
langgraph==0.4.1
langchain==0.3.25
langchain-openai==0.3.14
langchain-tavily==0.1.6
python-dotenv==1.1.0
pydantic==2.11.1
pydantic-settings==2.8.0
httpx==0.27.2
tavily-python
reportlab==4.1.0
